% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd2030.extractor-package.R
\docType{package}
\name{cd2030.extractor-package}
\alias{cd2030.extractor}
\alias{cd2030.extractor-package}
\title{cd2030.extractor: CD2030 Data Extractor}
\description{
A Shiny application for extracting and managing data from DHIS2. It provides an interactive interface to query, filter, and export DHIS2 data for further analysis. Designed for efficiency and ease of use in data-driven decision-making.
}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com}

Authors:
\itemize{
  \item Martin Mutua
  \item Peter Kaberia \email{pkaberia@aphrc.org}
  \item Godfrey Adero \email{gadero@aphrc.org}
  \item Rornald Muhumuza \email{rmuhumuza@aphrc.org}
}

Other contributors:
\itemize{
  \item African Population and Health Research Center (APHRC) [copyright holder, funder]
}

}
\keyword{internal}
